backend/requirements.txt
pandas==2.3.3
openai==2.20.0
chromadb==1.5.2
python-dotenv==1.2.1
langchain==1.2.10
langchain-openai==1.1.10
langchain-community==0.4.1
langgraph==1.0.10
langsmith==0.7.11

fastapi==0.135.1
uvicorn[standard]==0.41.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==5.0.0
python-multipart==0.0.22
